restructure kms primitives
